refactor(cjs): resolver organized in multiple files
