+++ src/esm/api/register.ts
@@ -2,7 +2,7 @@ import module from 'node:module';
 import { MessageChannel, type MessagePort } from 'node:worker_threads';
 import type { Message } from '../types.js';
 import type { RequiredProperty } from '../../types.js';
-import { interopCjsExports } from '../../cjs/api/module-resolve-filename.js';
+import { interopCjsExports } from '../../cjs/api/module-resolve-filename/interop-cjs-exports.js';
 import { createScopedImport, type ScopedImport } from './scoped-import.js';
 
 export type TsconfigOptions = false | string;
